Move remaining loading/importing code to new MVC structure.
